Add optional all-sample MSD path to align GPUMD output with full-trajectory post-processing
Agent-Logs-Url: https://github.com/zhyan0603/GPUMD/sessions/b872327f-69de-4767-8b58-3f55ec9e5a7e

+++ src/measure/msd.cuh
@@ -71,7 +71,7 @@ private:
   int num_groups_;
   std::vector<int> num_atoms_per_group_;
   std::vector<int> group_per_atom_cpu_;
-  int num_time_origins_;
+  long long num_time_origins_;
   bool use_all_samples_ = false;
   std::vector<long long> num_time_origins_per_lag_;
   int save_output_every_ = -1;
